Programs to check server functionality
These additional programs are expected to be used at customer
installations to verify that the 'libkmip' library is compatible
with their KMIP server.

The demonstration "get" program was changed to be more customer
friendly by switching to 'getopt'.  Similarly, the "destroy"
was converted to "get" functionality to better understand the
limitations of the high-level API.